requirements.txt
numpy==1.24.3
pandas==2.0.1
pandas==0.19.2
requests==2.31.0
selenium==4.9.1
google-auth==2.19.0
